+++ linkerd/protocol/h2/src/main/scala/io/buoyant/linkerd/protocol/H2Config.scala
@@ -76,23 +76,19 @@ object H2Initializer extends H2Initializer
 class H2Config extends RouterConfig {
 
   var client: Option[H2Client] = None
+  var service: Option[H2Svc] = None
   var servers: Seq[H2ServerConfig] = Nil
 
   @JsonDeserialize(using = classOf[H2IdentifierConfigDeserializer])
   var identifier: Option[Seq[H2IdentifierConfig]] = None
 
   @JsonIgnore
-  override def baseResponseClassifier =
-    ResponseClassifiers.NonRetryableServerFailures
-      .orElse(super.baseResponseClassifier)
-
-  // TODO: gRPC (trailers-aware)
-  @JsonIgnore
-  override def responseClassifier =
-    ResponseClassifiers.NonRetryableStream(super.responseClassifier)
-
-  @JsonIgnore
   override val protocol: ProtocolInitializer = H2Initializer
+
+  @JsonIgnore
+  override val defaultResponseClassifier = ResponseClassifiers.NonRetryableStream(
+    ResponseClassifiers.NonRetryableServerFailures orElse ClassifiedRetries.Default
+  )
 
   @JsonIgnore
   override def routerParams: Stack.Params =
@@ -149,6 +145,38 @@ trait H2ClientConfig extends ClientConfig with H2EndpointConfig {
   @JsonIgnore
   override def params(vars: Map[String, String]) = withEndpointParams(super.params(vars))
 
+}
+
+@JsonTypeInfo(
+  use = JsonTypeInfo.Id.NAME,
+  include = JsonTypeInfo.As.EXISTING_PROPERTY,
+  property = "kind",
+  visible = true,
+  defaultImpl = classOf[H2DefaultSvc]
+)
+@JsonSubTypes(Array(
+  new JsonSubTypes.Type(value = classOf[H2DefaultSvc], name = "io.l5d.global"),
+  new JsonSubTypes.Type(value = classOf[H2StaticSvc], name = "io.l5d.static")
+))
+abstract class H2Svc extends Svc
+
+class H2DefaultSvc extends H2Svc with DefaultSvc with H2SvcConfig
+
+class H2StaticSvc(val configs: Seq[H2SvcPrefixConfig]) extends H2Svc with StaticSvc
+
+class H2SvcPrefixConfig(prefix: PathMatcher) extends SvcPrefixConfig(prefix) with H2SvcConfig
+
+trait H2SvcConfig extends SvcConfig {
+
+  @JsonIgnore
+  override def baseResponseClassifier =
+    ResponseClassifiers.NonRetryableServerFailures
+      .orElse(super.baseResponseClassifier)
+
+  // TODO: gRPC (trailers-aware)
+  @JsonIgnore
+  override def responseClassifier =
+    super.responseClassifier.map(ResponseClassifiers.NonRetryableStream(_))
 }
 
 class H2ServerConfig extends ServerConfig with H2EndpointConfig {