+++ linkerd/protocol/thrift/src/main/scala/io/buoyant/linkerd/protocol/ThriftInitializer.scala
@@ -44,6 +44,8 @@ case class ThriftConfig(
 ) extends RouterConfig {
 
   var servers: Seq[ThriftServerConfig] = Nil
+  var service: Option[Svc] = None
+
   @JsonProperty("client")
   var _client: Option[ThriftClient] = None