+++ CHANGES.md
@@ -3,8 +3,6 @@
 * Add support for per-client configuration.
 * Simplify TLS configuration.
 * Allow dtab fallback when consul returns an empty address set.
-* Move timeoutMs from router section to client section in linkerd config.
-* Add timeoutMs option to server section in linkerd config.
 * Automatically upgrade all HTTP/1.0 messages to HTTP/1.1.
 * Add `io.l5d.influxdb` LINE telemeter
 * Refactor http and h2 identifiers for consistency:
@@ -12,6 +10,9 @@
   * The `io.l5d.headerPath` id has been renamed to `io.l5d.header.path`.
   * The `io.l5d.h2.ingress` id has been renamed to `io.l5d.ingress`.
   * The `io.l5d.http.ingress` id has been renamed to `io.l5d.ingress`.
+* Add support for per-service configuration.
+* Split the timeoutMs router option into a requestAttemptTimeoutMs client option
+  and a totalTimeoutMs service option.
 
 ## 0.9.1 2017-03-15